+++ libubx/ubx.c
@@ -619,7 +619,7 @@ ubx_type_t *ubx_type_get_by_hash(ubx_node_info_t *ni, const uint8_t *hash)
 
 	HASH_ITER(hh, ni->types, type, tmptype) {
 		if (strncmp((char *)type->hash,
-			    (char *)hash, TYPE_HASH_LEN) == 0)
+			    (char *)hash, UBX_TYPE_HASH_LEN) == 0)
 			return type;
 	}
 	return NULL;
@@ -637,23 +637,23 @@ ubx_type_t *ubx_type_get_by_hashstr(ubx_node_info_t *ni, const char *hashstr)
 {
 	int len;
 	char tmp[3];
-	uint8_t hash[TYPE_HASH_LEN+1];
+	uint8_t hash[UBX_TYPE_HASH_LEN+1];
 
 	len = strlen(hashstr);
 
-	if (len != TYPE_HASHSTR_LEN) {
+	if (len != UBX_TYPE_HASHSTR_LEN) {
 		logf_err(ni, "invalid length of hashstr %u", len);
 		return NULL;
 	}
 
 	/* convert from string to binary array */
-	for (int i = 0; i < TYPE_HASH_LEN; i++) {
+	for (int i = 0; i < UBX_TYPE_HASH_LEN; i++) {
 		tmp[0] = hashstr[i * 2];
 		tmp[1] = hashstr[i * 2 + 1];
 		tmp[2] = '\0';
 		hash[i] = (uint8_t) strtoul(tmp, NULL, 16);
 	}
-	hash[TYPE_HASH_LEN] = '\0';
+	hash[UBX_TYPE_HASH_LEN] = '\0';
 
 	return ubx_type_get_by_hash(ni, hash);
 }
@@ -669,9 +669,9 @@ ubx_type_t *ubx_type_get_by_hashstr(ubx_node_info_t *ni, const char *hashstr)
  */
 void ubx_type_hashstr(const ubx_type_t *t, char *buf)
 {
-	for (int i = 0; i < TYPE_HASH_LEN; i++)
+	for (int i = 0; i < UBX_TYPE_HASH_LEN; i++)
 		sprintf(buf + i * 2, "%02x", t->hash[i]);
-	buf[TYPE_HASH_LEN * 2] = '\0';
+	buf[UBX_TYPE_HASH_LEN * 2] = '\0';
 }