+++ std_types/stdtypes/types/tstat.h
@@ -3,7 +3,7 @@
 
 struct ubx_tstat
 {
-	char block_name[BLOCK_NAME_MAXLEN+1];
+	char block_name[UBX_BLOCK_NAME_MAXLEN + 1];
 	struct ubx_timespec min;
 	struct ubx_timespec max;
 	struct ubx_timespec total;