+++ tests/test_type_lookup.lua
@@ -29,7 +29,7 @@ function TestTypeLookup:TestLookupIntByHash()
 end
 
 function TestTypeLookup:TestLookupIntByHashstr()
-   local t1_hstr = ffi.new("char["..tostring(ffi.C.TYPE_HASHSTR_LEN+1)..']')
+   local t1_hstr = ffi.new("char["..tostring(ffi.C.UBX_TYPE_HASHSTR_LEN+1)..']')
    local t1 = ubx.type_get(NI, "int")
    luaunit.assert_not_nil(t1, "failed to lookup type 'int'")
    ubx.type_hashstr(t1, t1_hstr)
@@ -45,14 +45,14 @@ function TestTypeLookup:TestLookupInvalidType()
 end
 
 function TestTypeLookup:TestLookupInvalidHash()
-   local hash = ffi.new("char["..tostring(ffi.C.TYPE_HASH_LEN+1)..']')
+   local hash = ffi.new("char["..tostring(ffi.C.UBX_TYPE_HASH_LEN+1)..']')
    local t1 = ubx.type_get_by_hash(NI, hash)
    luaunit.assert_nil(t1, "invalid type lookup succeded")
 end
 
 function TestTypeLookup:TestLookupInvalidHashstr()
-   local hashstr = ffi.new("char["..tostring(ffi.C.TYPE_HASHSTR_LEN+1)..']')
-   for i=0,ffi.C.TYPE_HASHSTR_LEN-1 do hashstr[i] = string.byte('a') end
+   local hashstr = ffi.new("char["..tostring(ffi.C.UBX_TYPE_HASHSTR_LEN+1)..']')
+   for i=0,ffi.C.UBX_TYPE_HASHSTR_LEN-1 do hashstr[i] = string.byte('a') end
    local t1 = ubx.type_get_by_hashstr(NI, hashstr)
    luaunit.assert_nil(t1, "invalid type lookup succeded")
 end